Save old changes.

+++ llt.py
@@ -185,8 +185,7 @@ class SimpleModel(object):
 #             0)
         '''FIXME: Assumes that sampling rate is constant, i.e. the sampling
         times are in arithmetic progression with rate self._tinter'''
-        tindex = min(
-            np.floor(indices[1]/self._tinter), self._lsignals - 1)
+        tindex = int(min(np.floor(indices[1]/self._tinter), self._lsignals - 1))
         # assert 0 <= tindex <= len(self._signals[-1]), \
         #        'Invalid query outside the time domain of the trace! %f' % tindex
         return self._signals[indices[0]][tindex]
